add custom button for cron job

+++ src/main/java/cz/jiripinkas/jba/repository/OTPRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import cz.jiripinkas.jba.entity.OTP;
 
-public interface OTPRepository extends JpaRepository<OTP, Long> {
+public interface OTPRepository extends JpaRepository<OTP, Integer> {
 
     OTP findByOtp(String otp);
