+++ src/main/java/cz/jiripinkas/jba/repository/VerificationTokenRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import cz.jiripinkas.jba.entity.User;
 import cz.jiripinkas.jba.entity.VerificationToken;
 
-public interface VerificationTokenRepository extends JpaRepository<VerificationToken, Long> {
+public interface VerificationTokenRepository extends JpaRepository<VerificationToken, Integer> {
 
     VerificationToken findByToken(String token);
